+++ src/js/edu-benefits/utils/educationStatus.js
@@ -1,5 +1,5 @@
 import EducationWizard from '../components/EducationWizard';
-import ApplicationStatus from '../../common/schemaform/ApplicationStatus';
+import ApplicationStatus from '../../common/schemaform/save-in-progress/ApplicationStatus';
 
 // helper module so that we can code split with both of these components
 // in the same bundle
