+++ test/common/schemaform/save-in-progress/FormSaved.unit.spec.jsx
@@ -3,7 +3,7 @@ import moment from 'moment';
 import { expect } from 'chai';
 import SkinDeep from 'skin-deep';
 import sinon from 'sinon';
-import { FormSaved } from '../../../src/js/common/schemaform/FormSaved';
+import { FormSaved } from '../../../../src/js/common/schemaform/save-in-progress/FormSaved';
 
 describe('Schemaform <FormSaved>', () => {
   let windowOpen;
