+++ test/common/schemaform/save-in-progress/FormStartControls.unit.spec.jsx
@@ -5,8 +5,8 @@ import SkinDeep from 'skin-deep';
 import sinon from 'sinon';
 import ReactTestUtils from 'react-dom/test-utils';
 
-import { getFormDOM } from '../../util/schemaform-utils';
-import { FormStartControls } from '../../../src/js/common/schemaform/FormStartControls';
+import { getFormDOM } from '../../../util/schemaform-utils';
+import { FormStartControls } from '../../../../src/js/common/schemaform/save-in-progress/FormStartControls';
 
 describe('Schemaform <FormStartControls>', () => {
   const startPage = 'testing';
